Ammends bib environment & Glossa pub
This commit ammends the bibliography environment in order to print the
pubstate in the leftmargin as a label.

This commit also kills the addendum+pubstate bibmacro since the pubstate
is being printed in the left margin.

This commit also adds the forthcoming Glossa special issue publication
as part of the conference proceedings, and there are some minor changes
to the .tex file

I still need to fix the sorting scheme, but this is fine for now since
there's only one publication.

Unrelated to the bibliography environment, but I also still need to fix
the fact that my name is not bolded in the references.